agentic AI/requirements.txt
# Agentic AI Service for Retail Management
# Requires: Python 3.10+

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.13
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
